requirements.txt
torch==1.6
torchvision==0.7.0
tensorboard==2.3.0
pillow
pyyaml>=5.3
progressbar2
numpy